Fix some bugs, and add shaders for cubes.

+++ data/shaders/cube.frag.c
@@ -2,7 +2,7 @@
 
 uniform sampler2D tex;
 
-varying int select;
+varying float select;
 
 void main()
 {
